refactor: move functions
Move upm-config related functions to own module
